Update README.md
Overlord is a journey - a journey of transformation, metamorphosis and identity all wrapped up in the isekai format. As usual we have an overpowered protagonist that finds himself in a fantasy world. But this time he is reincarnated as the character, Momonga, he created in his favourite MMO. This is key because he lived out his Edge Lord fantasies online by playing as a non-human evil Overlord. So, now he really is an overlord called Ains Owl Gown, and must live up to his title in all the ways an Overlord should. That means all the ruthlessness, apathy and destructiveness as well as power.